refactor: [#91] service & configuration
