+++ waflog/src/main/kotlin/com/wafflestudio/waflog/global/config/AWSConfig.kt
@@ -1,0 +1,28 @@
+package com.wafflestudio.waflog.global.config
+
+import com.amazonaws.auth.AWSStaticCredentialsProvider
+import com.amazonaws.auth.BasicAWSCredentials
+import com.amazonaws.regions.Regions
+import com.amazonaws.services.s3.AmazonS3
+import com.amazonaws.services.s3.AmazonS3ClientBuilder
+import org.springframework.beans.factory.annotation.Value
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+
+@Configuration
+class AWSConfig {
+    @Bean
+    fun assetS3Client(
+        @Value("\${cloud.aws.credentials.access-key}")
+        accessKey: String,
+        @Value("\${cloud.aws.credentials.secret-key}")
+        secretKey: String
+
+    ): AmazonS3 {
+        return AmazonS3ClientBuilder
+            .standard()
+            .withCredentials(AWSStaticCredentialsProvider(BasicAWSCredentials(accessKey, secretKey)))
+            .withRegion(Regions.AP_NORTHEAST_2)
+            .build()
+    }
+}
