+++ waflog/src/main/resources/application.yml
@@ -31,7 +31,7 @@ spring:
   jpa:
     generate-ddl: true
     hibernate:
-      ddl-auto: update
+      ddl-auto: create
 
 ---
 spring:
